+++ ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vintran <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: vintran <vintran@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 12:32:52 by vintran           #+#    #+#             */
-/*   Updated: 2021/01/04 12:32:58 by vintran          ###   ########.fr       */
+/*   Updated: 2021/12/17 13:56:05 by vintran          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t		ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int	i;
 
 	i = 0;
 	while (s[i])
