+++ ft_strnstr.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnstr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vintran <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: vintran <vintran@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 12:33:40 by vintran           #+#    #+#             */
-/*   Updated: 2021/01/08 09:10:23 by vintran          ###   ########.fr       */
+/*   Updated: 2021/12/17 13:56:24 by vintran          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char		*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, int len)
 {
-	size_t i;
-	size_t j;
+	int	i;
+	int	j;
 
 	i = 0;
 	if (!needle[0])
@@ -23,8 +23,8 @@ char		*ft_strnstr(const char *haystack, const char *needle, size_t len)
 	while (haystack[i] && i < len)
 	{
 		j = 0;
-		while (haystack[i + j] && needle[j] &&
-				haystack[i + j] == needle[j] && i + j < len)
+		while (haystack[i + j] && needle[j]
+			&& haystack[i + j] == needle[j] && i + j < len)
 			j++;
 		if (!needle[j])
 			return ((char *)&haystack[i]);
