+++ libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vintran <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: vintran <vintran@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 12:35:52 by vintran           #+#    #+#             */
-/*   Updated: 2021/01/04 12:36:04 by vintran          ###   ########.fr       */
+/*   Updated: 2021/12/17 14:01:27 by vintran          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ struct		s_list
 };
 
 int			ft_atoi(const char *str);
-void		ft_bzero(void *s, size_t n);
-void		*ft_calloc(size_t nmemb, size_t size);
+void		ft_bzero(void *s, int n);
+void		*ft_calloc(int nmemb, int size);
 int			ft_isalnum(int c);
 int			ft_isalpha(int c);
 int			ft_isascii(int c);
@@ -42,12 +42,12 @@ t_list		*ft_lstlast(t_list *lst);
 t_list		*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
 t_list		*ft_lstnew(void *content);
 int			ft_lstsize(t_list *lst);
-void		*ft_memccpy(void *dst, const void *src, int c, size_t n);
-void		*ft_memchr(const void *s, int c, size_t n);
-int			ft_memcmp(const void *s1, const void *s2, size_t n);
-void		*ft_memcpy(void *dst, const void *src, size_t n);
-void		*ft_memmove(void *dst, const void *src, size_t len);
-void		*ft_memset(void *b, int c, size_t len);
+void		*ft_memccpy(void *dst, const void *src, int c, int n);
+void		*ft_memchr(const void *s, int c, int n);
+int			ft_memcmp(const void *s1, const void *s2, int n);
+void		*ft_memcpy(void *dst, const void *src, int n);
+void		*ft_memmove(void *dst, const void *src, int len);
+void		*ft_memset(void *b, int c, int len);
 void		ft_putchar_fd(char c, int fd);
 void		ft_putendl_fd(char *s, int fd);
 void		ft_putnbr_fd(int n, int fd);
@@ -56,15 +56,15 @@ char		**ft_split(const char *s, char c);
 char		*ft_strchr(const char *s, int c);
 char		*ft_strdup(const char *s);
 char		*ft_strjoin(char const *s1, char const *s2);
-size_t		ft_strlcat(char *dst, const char *src, size_t size);
-size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize);
-size_t		ft_strlen(const char *s);
+int			ft_strlcat(char *dst, const char *src, int size);
+int			ft_strlcpy(char *dst, const char *src, int dstsize);
+int			ft_strlen(const char *s);
 char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-int			ft_strncmp(const char *s1, const char *s2, size_t n);
-char		*ft_strnstr(const char *s1, const char *s2, size_t n);
+int			ft_strncmp(const char *s1, const char *s2, int n);
+char		*ft_strnstr(const char *s1, const char *s2, int n);
 char		*ft_strrchr(const char *s, int c);
 char		*ft_strtrim(char const *s1, char const *set);
-char		*ft_substr(char const *s, unsigned int start, size_t len);
+char		*ft_substr(char const *s, unsigned int start, int len);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
